requirements.txt
Flask==2.0.1
flask-session=0.4.0
flask-ext=0.1
flask-markdown=0.3
flask-pymongo=2.3
bcrypt=3.2.0
jwt=1.2.0